--- a/benchmarks/time_utils.h
+++ b/benchmarks/time_utils.h
@@ -21,9 +21,10 @@ using namespace gtn;
 
 #define TIME_DEVICE(FUNC, DEVICE) \
   { \
-    auto deviceName = Device(DEVICE).isCuda() ? "(cuda)" : "(cpu)"; \
+    bool isCuda = Device(DEVICE).isCuda(); \
+    auto deviceName = isCuda ? "(cuda)" : "(cpu)"; \
     std::cout << "Timing " << #FUNC << " " << deviceName << " ...  " << std::flush; \
-    std::cout << std::setprecision(5) << timeit(FUNC, true) << " msec" << std::endl; \
+    std::cout << std::setprecision(5) << timeit(FUNC, isCuda) << " msec" << std::endl; \
   }
 
 #define TIME(FUNC) \