--- a/gtn/cpu/shortest.cpp
+++ b/gtn/cpu/shortest.cpp
@@ -8,7 +8,7 @@
 #include <algorithm>
 #include <cmath>
 #include <limits>
-#include <queue>
+#include <stack>
 
 #include "gtn/cpu/shortest.h"
 
@@ -36,65 +36,80 @@ void shortestDistanceGrad(
     const Graph& deltas,
     const std::vector<int>& sortedNodes,
     const std::vector<float>& nodeScores,
-    const std::vector<float>& maxScoresCache,
     const std::vector<size_t>& maxArcIdxCache,
     bool tropical) {
 
   std::vector<float> nodeGrads(g.numNodes(), 0.0);
   std::vector<float> arcGrads(g.numArcs(), 0.0);
-  float curScore = 0.0;
-  float denom = tropical ? 0.0f : std::exp(output - maxScoresCache.back());
-  for (auto n : g.accept()) {
-    if (tropical) {
-      curScore = (n == maxArcIdxCache.back()) ? 1.0f : 0.0f;
-    } else if (maxScoresCache.back() == kNegInf) {
-      curScore = 0.0f;
-    } else {
-      curScore = std::exp(nodeScores[n] - maxScoresCache.back()) / denom;
-    }
-    nodeGrads[n] += curScore;
-  }
 
   for (int i = sortedNodes.size() - 1; i >= 0; --i) {
     auto n = sortedNodes[i];
-    denom = tropical ? 0.0f : std::exp(nodeScores[n] - maxScoresCache[n]);
-    for (const auto a : g.in(n)) {
-      auto un = g.srcNode(a);
+    for (const auto a : g.out(n)) {
+      auto dn = g.dstNode(a);
+      float curScore;
       if (tropical) {
-        curScore = (a == maxArcIdxCache[n]) ? nodeGrads[n] : 0.0f;
-      } else if (maxScoresCache[n] == kNegInf) {
+        curScore = (a == maxArcIdxCache[dn]) ? nodeGrads[dn] : 0.0f;
+      } else if (nodeScores[dn] == kNegInf) {
         curScore = 0.0f;
       } else {
-        curScore = nodeGrads[n] *
-            std::exp(nodeScores[un] + g.weight(a) - maxScoresCache[n]) / denom;
+        curScore = nodeGrads[dn] *
+            std::exp(nodeScores[n] + g.weight(a) - nodeScores[dn]);
       }
-      nodeGrads[un] += curScore;
-      arcGrads[a] = curScore * deltas.item();
+      nodeGrads[n] += curScore;
+      arcGrads[a] = curScore;
+    }
+    if (g.isAccept(n)) {
+      if (tropical) {
+        nodeGrads[n] += (n == maxArcIdxCache.back()) ? deltas.item() : 0.0f;
+      } else if (nodeScores.back() != kNegInf) {
+        nodeGrads[n] += std::exp(nodeScores[n] - output) * deltas.item();
+      }
     }
   }
   g.addGrad(std::move(arcGrads));
 }
 
 std::vector<int> topSort(const Graph& g) {
-  std::vector<int> degrees(g.numNodes());
+  std::stack<std::tuple<int, int, int>> nodeStack;
+  std::vector<bool> visited(g.numNodes(), false);
   std::vector<int> sortedNodes(g.numNodes());
-  int start = 0;
-  int end = 0;
+
+  int numVisited = 0;
   for (int n = 0; n < g.numNodes(); ++n) {
-    degrees[n] = g.numIn(n);
-    if (degrees[n] == 0) {
-      sortedNodes[end++] = n;
+    if (g.numIn(n) == 0) {
+      nodeStack.emplace(n, 0, 0);
+      visited[n] = true;
+      numVisited++;
     }
   }
-  while (start < g.numNodes()) {
-    int oldEnd = end;
-    while (start < oldEnd) {
-      for (auto a : g.out(sortedNodes[start++])) {
-        auto dst = g.dstNode(a);
-        if (--degrees[dst] == 0) {
-          sortedNodes[end++] = dst;
-        }
+
+  int diameter = 0;
+  int nodeIdx = g.numNodes() - 1;
+  while (!nodeStack.empty()) {
+    int n, a, d;
+    std::tie(n, a, d) = nodeStack.top();
+    diameter = std::max(diameter, d);
+    nodeStack.pop();
+
+    // Stop early if all the nodes have been visited already
+    if (numVisited == g.numNodes()) {
+      sortedNodes[nodeIdx--] = n;
+      continue;
+    }
+
+    auto it = g.out(n).begin() + a;
+    for (; it < g.out(n).end(); it++) {
+      auto dst = g.dstNode(*it);
+      if (!visited[dst]) {
+        nodeStack.emplace(n, it - g.out(n).begin(), d);
+        nodeStack.emplace(dst, 0, d + 1);
+        visited[dst] = true;
+        numVisited++;
+        break;
       }
+    }
+    if (it == g.out(n).end()) {
+      sortedNodes[nodeIdx--] = n;
     }
   }
   return sortedNodes;
@@ -103,9 +118,7 @@ std::vector<int> topSort(const Graph& g) {
 } // namespace
 
 Graph shortestDistance(const Graph& g, bool tropical /* = false */) {
-  // List of scores for each node
   std::vector<float> scores(g.numNodes());
-  std::vector<float> maxScoresCache(g.numNodes() + 1, kNegInf);
   std::vector<size_t> maxArcIdxCache(g.numNodes() + 1, -1);
   auto sortedNodes = topSort(g);
 
@@ -123,49 +136,48 @@ Graph shortestDistance(const Graph& g, bool tropical /* = false */) {
     return maxScore + std::log1p(score);
   };
 
-  std::vector<float> inScores;
-  float maxScore = kNegInf;
   for (auto n : sortedNodes) {
+    float maxScore = kNegInf;
+    std::vector<float> inScores(g.numIn(n) + g.isStart(n));
+    int i = 0;
     for (auto a : g.in(n)) {
       auto un = g.srcNode(a);
-      inScores.push_back(scores[un] + g.weight(a));
-      if (inScores.back() > maxScoresCache[n]) {
-        maxScoresCache[n] = inScores.back();
+      inScores[i] = scores[un] + g.weight(a);
+      if (inScores[i]  > maxScore) {
+        maxScore = inScores[i];
         maxArcIdxCache[n] = a;
       }
+      i++;
     }
     if (g.isStart(n)) {
-      inScores.push_back(0.0);
-      if (inScores.back() > maxScoresCache[n]) {
-        maxScoresCache[n] = inScores.back();
+      inScores[i] = 0.0;
+      if (inScores[i] > maxScore) {
+        maxScore = inScores[i];
         maxArcIdxCache[n] = -1; // an invalid value
       }
     }
-    scores[n] = getScore(inScores, maxScoresCache[n]);
-    inScores.clear();
-    maxScore = kNegInf;
+    scores[n] = getScore(inScores, maxScore);
   }
 
   // Accumulate scores at all the accept nodes.
+  std::vector<float> inScores;
+  float maxScore = kNegInf;
   for (auto n : g.accept()) {
     inScores.push_back(scores[n]);
-    if (inScores.back() > maxScoresCache.back()) {
-      maxScoresCache.back() = std::max(maxScoresCache.back(), inScores.back());
+    if (inScores.back() > maxScore) {
+      maxScore = inScores.back();
       maxArcIdxCache.back() = n; // NOTE: Using node idx (instead of arc idx)
     }
   }
-  auto score = getScore(inScores, maxScoresCache.back());
+  auto score = getScore(inScores, maxScore);
 
   // clear cache not required for bwd
-  if (tropical) {
-    maxScoresCache.clear();
-  } else {
+  if (!tropical) {
     maxArcIdxCache.clear();
   }
 
   auto gradFunc = [scores = std::move(scores),
                    sortedNodes = std::move(sortedNodes),
-                   maxScoresCache = std::move(maxScoresCache),
                    maxArcIdxCache = std::move(maxArcIdxCache),
                    output = score,
                    tropical](std::vector<Graph>& inputs, Graph deltas) mutable {
@@ -175,7 +187,6 @@ Graph shortestDistance(const Graph& g, bool tropical /* = false */) {
         deltas,
         sortedNodes,
         scores,
-        maxScoresCache,
         maxArcIdxCache,
         tropical);
   };
@@ -188,36 +199,23 @@ Graph shortestDistance(const Graph& g, bool tropical /* = false */) {
 }
 
 Graph shortestPath(const Graph& g) {
-  std::queue<int> computed;
-  // List of in degrees for each node
-  std::vector<size_t> degrees(g.numNodes());
   // List of scores and backpointers for each node
   std::vector<int> backPointers(g.numNodes());
   std::vector<float> scores(g.numNodes(), kNegInf);
-  for (size_t n = 0; n < g.numNodes(); ++n) {
-    degrees[n] = g.numIn(n);
-  }
+  auto sortedNodes = topSort(g);
+
   for (auto n : g.start()) {
     scores[n] = 0.0;
     backPointers[n] = -1;
-    if (g.numIn(n) == 0) {
-      computed.push(n);
-    }
   }
 
-  while (!computed.empty()) {
-    auto n = computed.front();
-    computed.pop();
-    auto score = scores[n];
-    for (auto a : g.out(n)) {
-      auto dn = g.dstNode(a);
-      auto nScore = score + g.weight(a);
-      if (nScore > scores[dn]) {
-        scores[dn] = nScore;
-        backPointers[dn] = a;
-      }
-      if ((--degrees[dn]) == 0) {
-        computed.push(dn);
+  for (auto n : sortedNodes) {
+    for (auto a : g.in(n)) {
+      auto un = g.srcNode(a);
+      auto nScore = scores[un] + g.weight(a);
+      if (nScore > scores[n]) {
+        scores[n] = nScore;
+        backPointers[n] = a;
       }
     }
   }
@@ -226,10 +224,6 @@ Graph shortestPath(const Graph& g) {
   float score = kNegInf;
   int best = -1;
   for (auto a : g.accept()) {
-    if (degrees[a] > 0) {
-      throw std::invalid_argument(
-          "Graph has a cycle, self-loop or is disconnected!");
-    }
     if (scores[a] > score) {
       score = scores[a];
       best = a;