--- a/gtn/graph.cpp
+++ b/gtn/graph.cpp
@@ -223,7 +223,7 @@ void Graph::addGrad(const Graph& other) {
   if (device() != other.device()) {
     throw std::invalid_argument("[Graph::addGrad] device mismach");
   }
-  if (calcGrad() & other.numArcs() != numArcs()) {
+  if (calcGrad() && other.numArcs() != numArcs()) {
     throw std::logic_error("[Graph::addGrad] Invalid grad size.");
   }
   addGrad(other.weights());
@@ -311,6 +311,89 @@ void Graph::arcSort(bool olabel /* = false */) {
         sharedGraph_->outArcs.begin() + end,
         sortFn);
   }
+}
+
+void Graph::topSort() {
+  if (sharedGraph_->topSorted) {
+    return;
+  }
+  sharedGraph_->topSorted = true;
+  std::vector<int> degrees(numNodes());
+  std::vector<int> sortedNodes(numNodes());
+  std::vector<int> widths;
+  int start = 0;
+  int end = 0;
+  for (int n = 0; n < numNodes(); ++n) {
+    degrees[n] = numIn(n);
+    if (degrees[n] == 0) {
+      sortedNodes[end++] = n;
+    }
+  }
+  while (start < numNodes()) {
+    widths.push_back(end - start);
+    int oldEnd = end;
+    while (start < oldEnd) {
+      for (auto a : out(sortedNodes[start++])) {
+        auto dst = dstNode(a);
+        if (--degrees[dst] == 0) {
+          sortedNodes[end++] = dst;
+        }
+      }
+    }
+  }
+
+  HDSpan<int> startIds;
+  startIds.reserve(numStart());
+  HDSpan<int> acceptIds;
+  acceptIds.reserve(numAccept());
+  HDSpan<bool> start_(numNodes());
+  HDSpan<bool> accept_(numNodes());
+  HDSpan<int> inArcOffset(numNodes() + 1);
+  HDSpan<int> outArcOffset(numNodes() + 1);
+  inArcOffset[0] = 0;
+  outArcOffset[0] = 0;
+  HDSpan<int> inArcs(numArcs());
+  HDSpan<int> outArcs(numArcs());
+  HDSpan<int> srcNodes(numArcs());
+  HDSpan<int> dstNodes(numArcs());
+  for (int n = 0; n < numNodes(); ++n) {
+    auto oldN = sortedNodes[n];
+    if (isStart(oldN)) {
+      startIds.push_back(n);
+      start_[n] = true;
+    }
+    if (isAccept(oldN)) {
+      acceptIds.push_back(n);
+      accept_[n] = true;
+    }
+    inArcOffset[n + 1] = inArcOffset[n] + numIn(oldN);
+    outArcOffset[n + 1] = outArcOffset[n] + numOut(oldN);
+    std::copy(
+        in(oldN).begin(),
+        in(oldN).end(),
+        inArcs.begin() + inArcOffset[n]);
+    std::copy(
+        out(oldN).begin(),
+        out(oldN).end(),
+        outArcs.begin() + outArcOffset[n]);
+    for (auto a : in(oldN)) {
+      dstNodes[a] = n;
+    }
+    for (auto a : out(oldN)) {
+      srcNodes[a] = n;
+    }
+  }
+  auto& data = getData();
+  data.start = std::move(start_);
+  data.startIds = std::move(startIds);
+  data.accept = std::move(accept_);
+  data.acceptIds = std::move(acceptIds);
+  data.inArcOffset = std::move(inArcOffset);
+  data.outArcOffset = std::move(outArcOffset);
+  data.inArcs = std::move(inArcs);
+  data.outArcs = std::move(outArcs);
+  data.srcNodes = std::move(srcNodes);
+  data.dstNodes = std::move(dstNodes);
 }
 
 void Graph::setWeights(const float* weights) {