TImings on my laptop using `./benchmarks/benchmark_functions`

**After**
Timing forwardScoreLinearForward (cpu) ...  9.1351 msec
Timing forwardScoreLinearBackward (cpu) ...  9.085 msec
Timing forwardScoreRandDAGForward (cpu) ...  5.7167 msec
Timing forwardScoreRandDAGBackward (cpu) ...  6.4635 msec
Timing viterbiPathLinearForward (cpu) ...  1.7872 msec
Timing viterbiPathLinearBackward (cpu) ...  1.7198 msec
Timing viterbiPathRandDAGForward (cpu) ...  2.0626 msec
Timing viterbiPathRandDAGBackward (cpu) ...  0.1356 msec

**Before**
Timing forwardScoreLinearForward (cpu) ...  14.57 msec
Timing forwardScoreLinearBackward (cpu) ...  16.383 msec
Timing forwardScoreRandDAGForward (cpu) ...  5.8892 msec
Timing forwardScoreRandDAGBackward (cpu) ...  11.59 msec
Timing viterbiPathLinearForward (cpu) ...  4.2249 msec
Timing viterbiPathLinearBackward (cpu) ...  1.7191 msec
Timing viterbiPathRandDAGForward (cpu) ...  1.7667 msec
Timing viterbiPathRandDAGBackward (cpu) ...  0.087886 msec